requirements.txt
aiohttp==3.8.1
ecdsa==0.18.0
hdwallets==0.1.2
mnemonic==0.20
requests==2.28.1
ujson==5.4.0
urllib3==1.26.12
websocket-client==1.3.3

bech32~=1.2.0
setuptools~=60.10.0